modif: install sdl3_ttf command has been added.

+++ src/types/common.zig
@@ -47,7 +47,7 @@ pub const Dimensions = struct {
     h: f32,
     d: f32,
 
-    pub fn scale(self: *Dimensions, s: f16) Dimensions {
+    pub fn scale(self: *Dimensions, s: f32) Dimensions {
         self.w *= s;
         self.h *= s;
         self.d *= s;
